Add files via upload
Abstract
Quantum tunnelling is a well-established process, and yet, a continually evolving area of
research. This report presents numerical studies of quantum tunnelling induced by strong
laser fields that distort the Coulomb potential barrier, enabling ionisation. This is a fun-
damental process in attosecond science, which seeks to observe electron dynamics in real
time. Surprisingly, we identifed an ionisation event occurring when the electric field is zero,
meaning there is no barrier present challenging the conventional understanding of quantum
tunnelling. However, detecting this is difficult, as this ionisation event’s yield is extremely
low. To gain a deeper understanding of the mechanisms underlying this rare process, we
employ a topological saddle point analysis and bifurcation theory. Although this approach
does not directly enhance the yield of this specific ionisation event (where tunnelling occurs
without a barrier) but it may reveal structural features that help identify where to look
for new or more detectable signatures. This is done by extending the electric field to two
dimensions and examining co-orthogonal field configurations, where we uncover a complex,
self-intersecting saddle point structure. This structure evolves as the mixing angle of the elec-
tric field transitions from bichromatic to a monochromatic configuration (θ= 90◦to θ= 0◦),
also referred to as the colour switchover, giving rise to fold and cusp bifurcations. These
bifurcations complicate the classification of quantum trajectories due to the intertwined na-
ture of the resulting saddle points. Using bifurcation theory, we propose a methodology for
classifying these intricate structures, and offer new insights into quantum tunnelling without
a barrier.