uploading numerical module report from year 2 bsc
Quantitative Insights: Exploring Physical Phenomena through Numerical Modelling
This project was part of the Intro to Numerical Modelling module at King's College London where I achieved one of the highest grades in the cohort of 83%.

The main objective of this report is to explore and numerically model physical phenomena, as well as optimizing the parameters impacting these models to visualize and solve these computational complex problems. This report developed a simulated annealing algorithm by gradually cooling the system through a temperature schedule function that found the global minimum path for 30 cities to be 2163. In addition, this report also modelled the Tacoma Bridge's incident, and further explored the relationship between the torsional angle θ and the vertical displacement Y, where the natural resonant frequency was found to be in between 0.13 to 0.16 Hz, and is 35% decreased in comparison to the literature value, ν = 0.2 (Enrera, 2022).